+++ src/test/resources/application.yml
@@ -1,4 +1,10 @@
 spring:
+  datasource:
+    url: jdbc:h2:tcp://localhost/~/hello;
+    username: sa
+    password:
+    driver-class-name: org.h2.Driver
+
   jpa:
     properties:
       hibernate:
